Avoid using any bundlers but use tsc instead (#192)
