+++ package.json
@@ -3,18 +3,18 @@
   "version": "4.0.0-beta.2",
   "description": "Decode JWT tokens, mostly useful for browser applications.",
   "type": "module",
-  "main": "build/cjs/jwt-decode.cjs",
-  "module": "build/esm/jwt-decode.js",
-  "types": "build/cjs/jwt-decode.d.cts",
+  "main": "build/cjs/index.js",
+  "module": "build/esm/index.js",
+  "types": "build/cjs/index.d.ts",
   "exports": {
     ".": {
       "import": {
-        "default": "./build/esm/jwt-decode.js",
-        "types": "./build/esm/jwt-decode.d.ts"
+        "default": "./build/esm/index.js",
+        "types": "./build/esm/index.d.ts"
       },
       "require": {
-        "default": "./build/cjs/jwt-decode.cjs",
-        "types": "./build/cjs/jwt-decode.d.cts"
+        "default": "./build/cjs/index.js",
+        "types": "./build/cjs/index.d.ts"
       }
     }
   },
@@ -29,11 +29,8 @@
   "url": "https://github.com/auth0/jwt-decode/issues",
   "homepage": "https://github.com/auth0/jwt-decode#readme",
   "scripts": {
-    "dev": "rollup --sourcemap --config --configPlugin typescript",
-    "build": "rimraf build && rollup --sourcemap --config --configPlugin typescript --environment NODE_ENV:production && npm run generate-types",
-    "generate-types": "npm run generate-types:es && npm run generate-types:cjs",
-    "generate-types:es": "tsup --entry.jwt-decode lib/index.ts --out-dir build/esm --dts-only --format esm",
-    "generate-types:cjs": "tsup --entry.jwt-decode lib/index.ts --out-dir build/cjs --dts-only --format cjs",
+    "prebuild": "rimraf build",
+    "build": "tsc -b ./tsconfig.cjs.json ./tsconfig.esm.json && echo '{\"type\": \"commonjs\"}'> build/cjs/package.json",
     "test": "npm run test:node && npm run test:browser",
     "test:node": "NODE_OPTIONS='--experimental-vm-modules --no-warnings' jest --coverage",
     "test:browser": "NODE_OPTIONS='--experimental-vm-modules --no-warnings' jest --coverage --testEnvironment=jsdom",
@@ -45,19 +42,12 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "^11.1.2",
     "jest": "^29.6.2",
     "jest-environment-jsdom": "^29.6.2",
     "jest-junit": "^16.0.0",
     "rimraf": "^5.0.1",
-    "rollup": "^3.27.2",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-serve": "^2.0.2",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
-    "tslib": "^2.6.1",
-    "tsup": "^7.2.0",
     "typescript": "^5.1.6"
   },
   "files": [
