+++ tsconfig.build.json
@@ -1,8 +1,8 @@
 {
   "extends": "./tsconfig.json",
+  "include": ["./lib"],
   "compilerOptions": {
     "target": "ES2017",
-    "rootDir": "./lib",
-    "sourceMap": true
+    "declaration": true
   }
 }
