+++ tsconfig.cjs.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.build.json",
+  "compilerOptions": {
+    "outDir": "./build/cjs",
+    "module": "CommonJS"
+  }
+}
