+++ tsconfig.esm.json
@@ -1,0 +1,6 @@
+{
+  "extends": "./tsconfig.build.json",
+  "compilerOptions": {
+    "outDir": "./build/esm"
+  }
+}
